Move warnings out of tools and make private
