+++ parcels/__init__.py
@@ -2,7 +2,7 @@ from ._version import version
 
 __version__ = version
 
-import warnings as _warnings
+import warnings as _stdlib_warnings
 
 from parcels.field import Field, VectorField
 from parcels.fieldset import FieldSet
@@ -13,11 +13,17 @@ from parcels.particle import (
     ParticleClass,
     Variable,
 )
+from parcels._warnings import (
+    FieldSetWarning,
+    FileWarning,
+    KernelWarning,
+    ParticleSetWarning,
+)
 from parcels.particlefile import ParticleFile
 from parcels.particleset import ParticleSet
 from parcels.tools import *
 
-_warnings.warn(
+_stdlib_warnings.warn(
     "This is an alpha version of Parcels v4. The API is not stable and may change without deprecation warnings.",
     UserWarning,
     stacklevel=2,
