+++ parcels/tools/__init__.py
@@ -3,4 +3,3 @@ from .exampledata_utils import *
 from .interpolation_utils import *
 from .loggers import *
 from .statuscodes import *
-from .warnings import *
